# Improve speed when retrieving an archive metadata