--- a/lib/archive-hooks/zip.js
+++ b/lib/archive-hooks/zip.js
@@ -63,19 +63,19 @@ exports.getEntries = (archive) => {
  * @public
  *
  * @param {String} archive - archive path
+ * @param {String[]} entries - archive entries
  * @param {String} file - archive file
  * @fulfil {ReadableStream} file
  * @returns {Promise}
  *
  * @example
- * zip.extractFile('path/to/my.zip', 'my/file').then((stream) => {
+ * const entries = zip.getEntries('path/to/my.zip');
+ * zip.extractFile('path/to/my.zip', entries, 'my/file').then((stream) => {
  *   stream.pipe('...');
  * });
  */
-exports.extractFile = (archive, file) => {
+exports.extractFile = (archive, entries, file) => {
   return new Bluebird((resolve, reject) => {
-    const entries = exports.getEntries(archive);
-
     if (!_.find(entries, {
       name: file
     })) {