--- a/lib/archive.js
+++ b/lib/archive.js
@@ -46,19 +46,21 @@ const IMAGE_EXTENSIONS = _.reduce(supportedFileTypes, (accumulator, file) => {
  * @param {String} filePath - entry file path
  * @param {Object} options - options
  * @param {Object} options.hooks - archive hooks
+ * @param {Object[]} options.entries - archive entries
  * @param {*} [options.default] - entry default value
  * @fulfil {*} contents
  * @returns {Promise}
  *
  * extractEntryByPath('my/archive.zip', '_info/logo.svg', {
  *   hooks: { ... },
+ *   entries: [ ... ],
  *   default: ''
  * }).then((contents) => {
  *   console.log(contents);
  * });
  */
 const extractEntryByPath = (archive, filePath, options) => {
-  const fileEntry = _.find(options.hooks.getEntries(archive), (entry) => {
+  const fileEntry = _.find(options.entries, (entry) => {
     return _.chain(entry.name)
       .split('/')
       .tail()
@@ -70,7 +72,7 @@ const extractEntryByPath = (archive, filePath, options) => {
     return Bluebird.resolve(options.default);
   }
 
-  return options.hooks.extractFile(archive, fileEntry.name)
+  return options.hooks.extractFile(archive, options.entries, fileEntry.name)
     .then(rindle.extract);
 };
 
@@ -91,7 +93,7 @@ const extractEntryByPath = (archive, filePath, options) => {
  *   getEntries: (archive) => {
  *     return [ ..., ..., ... ];
  *   },
- *   extractFile: (archive, file) => {
+ *   extractFile: (archive, entries, file) => {
  *     ...
  *   }
  * }).then((image) => {
@@ -113,16 +115,19 @@ exports.extractImage = (archive, hooks) => {
   const imageEntry = _.first(imageEntries);
 
   return Bluebird.props({
-    imageStream: hooks.extractFile(archive, imageEntry.name),
+    imageStream: hooks.extractFile(archive, entries, imageEntry.name),
     logo: extractEntryByPath(archive, '_info/logo.svg', {
-      hooks
+      hooks,
+      entries
     }),
     bmap: extractEntryByPath(archive, '_info/image.bmap', {
-      hooks
+      hooks,
+      entries
     }),
     manifest: _.attempt(() => {
       return extractEntryByPath(archive, '_info/manifest.json', {
         hooks,
+        entries,
         default: '{}'
       }).then((manifest) => {
         try {