Currently, analysing an archive (e.g: a ZIP) and returning all its
metadata its an slow operation. After some investigation, this was
mainly caused by not reusing the information about available entries and
fetching this information over and over again.

With the following optimisations, the `.getEntries()` function is only
called once. The following benchmarkings were done on a
`Jessie_PiTFT28c.zip` image:

Before:

```
$ time node benchmark.js
node benchmark.js  0.36s user 5.16s system 99% cpu 5.568 total
```

After:

```
$ time node benchmark.js
node benchmark.js  0.35s user 1.03s system 100% cpu 1.373 total
```

Consider that this image doesn't contain extra metadata like a logo, or
a bmap file. If it would, the gains would be probably even larger.

Signed-off-by: Juan Cruz Viotti <jviotti@openmailbox.org>